requirements.txt
# Flask==2.3.2
# torch==2.4.0
# torchvision==0.19.0
# Pillow==10.0.0
# opencv-python==4.8.0.74
# scikit-image==0.21.0
# matplotlib==3.7.2
# numpy==2.0.1

Flask== 2.3.2
torch== 2.0.1
torchvision== 0.15.2
opencv-python== 4.8.0.76
numpy== 1.25.0
Pillow== 10.0.0
scikit-image== 0.21.0